requirements.txt
pandas
snscrape
requests
python-dotenv
vaderSentiment
transformers
torch
yfinance